Changes for review comments [HZ-2050]
